async-kit ak-massive stream wrapper for massive-js
massive-js supports streaming as an option on calls, especially to
SQL files that it loads up from disk.  However, it still uses a
callback style to do that.  This library adapts all the methods
loaded up from SQL files to return a stream that can be processed
just like any other stream.  This makes it easy to combine with
stream libraries like e.g. highlandjs or the other async-kit
components like ak-data.